Py_DECREF() item instead of Py_INCREF() after PyList_Append()
Py_BuildValue() sets the returned object's ref count to 1.
PyList_Append() auto increments the passed item's ref count, and
then Py_INCREF() was being called on the item.  This was making
the ref count 3, when we really want it to be 1.  We need to call
Py_DECREF() instead of Py_INCREF() to get the ref count to 1 or
it results in a memory leak as observed in

import pytricia
pyt = pytricia.PyTricia()
pyt['10.0.0.0/8'] = 1
while True:
   val = pyt.keys()